backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
reportlab==4.0.7
pdfplumber==0.11.4
zeep==4.2.1
lxml==4.9.3
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.13.1
psycopg2-binary==2.9.9
PyJWT==2.8.0


